+++ sdk/typescript/src/signature-mpc/encrypt_user_share.ts
